Arcane Explosion and SFX
Arcane Explosion is not complete and sound effects have been added for it, along with the OOM. Finished!

+++ Assets/Scripts/ManaBar_cs.cs
@@ -7,13 +7,13 @@ public class ManaBar_cs : MonoBehaviour
 {
     public Slider _manaSlider;
 
-    public void SetMaxMana(int mana)
+    public void SetMaxMana(float mana)
     {
         _manaSlider.maxValue = mana;
         _manaSlider.value = mana;
     }
 
-    public void SetMana(int mana)
+    public void SetMana(float mana)
     {
         _manaSlider.value = mana;
     }
